refactor: moved reset code to PlayScenarioPage chore: fixed type in post urls
