+++ frontend/src/containers/pages/PlayScenarioPage/PlayScenarioCanvas.jsx
@@ -1,8 +1,3 @@
-import { usePost } from "hooks/crudHooks";
-import { useContext } from "react";
-import { useParams } from "react-router-dom";
-import AuthenticationContext from "../../../context/AuthenticationContext";
-
 import CountdownTimer from "../../../components/TimerComponent";
 import componentResolver from "./componentResolver";
 
@@ -12,29 +7,14 @@ import componentResolver from "./componentResolver";
  * @component
  */
 
-export default function PlayScenarioCanvas({ scene, incrementor }) {
-  const { sceneId } = useParams();
-  const { user, getUserIdToken } = useContext(AuthenticationContext);
-
-  // Define the reset function inside the component
-  const reset = async () => {
-    const userId = user.uid;
-    const reqBody = { userId, currentScene: sceneId };
-    console.log("resetting");
-
-    await usePost(`api/navigate/group/reset/:groupId`, reqBody, getUserIdToken);
-    await usePost(`api/navigate/user/reset/:scenarioId`, reqBody, getUserIdToken);
-
-    console.log("reset");
-    // window.location.reload();
-  };
-
+export default function PlayScenarioCanvas({ scene, incrementor, reset }) {
   return (
     <>
       {scene.components?.map((component, index) => {
-        const action = component.type === "RESET_BUTTON"
-          ? reset
-          : () => component.nextScene && incrementor(component.nextScene);
+        const action =
+          component.type === "RESET_BUTTON"
+            ? reset
+            : () => component.nextScene && incrementor(component.nextScene);
 
         return componentResolver(component, index, action);
       })}
