+++ frontend/src/containers/pages/PlayScenarioPage/componentResolver.jsx
@@ -26,15 +26,15 @@ export default function componentResolver(component, index, onClick) {
           component={component}
         />
       );
-      case "RESET_BUTTON":
-        return (
-          <ButtonComponent
-            key={component.id}
-            id={index}
-            onClick={onClick}
-            component={component}
-          />
-        );
+    case "RESET_BUTTON":
+      return (
+        <ButtonComponent
+          key={component.id}
+          id={index}
+          onClick={onClick}
+          component={component}
+        />
+      );
     case "SPEECHTEXT":
     case "TEXT":
       return (
